fixup main rebase by bringing over index sort opto
